pyproject.toml
[project]
name = "snapopedia"
version = "0.1.0"
description = "Snapopedia backend service"
readme = "AGENTS.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.110.2",
    "uvicorn[standard]==0.29.0",
    "python-dotenv==1.0.1",
    "pydantic-settings==2.2.1",
    "httpx==0.27.0",
    "boto3==1.35.59",
    "python-multipart==0.0.9",
]

[project.optional-dependencies]
dev = [
    "pytest==8.2.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.hatch.build.targets.wheel.sources]
src = "src"

[tool.uv]
dev-dependencies = ["pytest==8.2.1"]
